Remove redundant navbar
